requirements.txt
# FastAPI 백엔드 서버 실행을 위한 추가 의존성
fastapi>=0.115.8
uvicorn>=0.34.0
websockets>=12.0
python-multipart>=0.0.9

# 기존 browser-use 의존성들
httpx>=0.27.2
pydantic>=2.10.4
python-dotenv>=1.0.1
requests>=2.32.3
posthog>=3.7.0
playwright>=1.49.0
setuptools>=75.8.0
markdownify==0.14.1
langchain-core>=0.3.35

# LLM 모델 의존성
langchain>=0.3.18
langchain-openai==0.3.1 